Update
Fixed the CRC-16 issue. The reaction wheel should be able to start and run manually using the "RW Manual Ctrl.vi". If the solution works, try connect the cosine function generator to the "control value" control inside the program and switch to speed mode so as to test the accelerating and decelerating ability of the wheel.